SpeechRecognizorActivity with abstract methods
