+++ app/src/main/java/SyntheseVocale/Interpretation.java
@@ -5,23 +5,30 @@ import android.content.Intent;
 
 import com.example.galidog2.ChoixMemorisationActivity;
 
+import java.util.ArrayList;
+
 public class Interpretation {
 
     private Activity activity;
     private VoiceOut voiceOut;
+    private ArrayList<String> matchList = new ArrayList<String>();
 
     public Interpretation(Activity activity) {
         this.activity = activity;
         voiceOut = new VoiceOut(activity.getApplicationContext());
     }
 
-    public void interpreterText(String text) {
-        if (text.contains("mémorisation")) {
-            Intent intent = new Intent(this.activity, ChoixMemorisationActivity.class);
-            this.activity.startActivity(intent);
+    public String findMatch(String text) {
+        addMatchs();
+        for (int i=0; i< matchList.size(); i++)
+            if (text.contains(matchList.get(i)))
+                return matchList.get(i);
+        return null;
+    }
 
-            this.voiceOut.speak("Mode mémorisation");
-        }
+    private void addMatchs(){
+        matchList.add("mémorisation");
+        matchList.add("ajouter trajet");
     }
 
 }
